Create README.md
Hand-made jewelry management system is an application developed on HTML and CSS for the user to know the basic working of a e-commerce site. It contains seven tables namely  register, user, paperquilling, silk, orders, gifts and review. The project consists of three stored procedures and a trigger. Tables in the database are linked at some point.

DBMS : A database management system (DBMS) is system software for creating and managing databases. The DBMS provides users and programmers with a systematic way to create, retrieve, update and manage data. A DBMS makes it possible for end users to create, read, update and delete data in a database. The DBMS essentially serves as an interface between the database and endusers or application programs, ensuring that data is consistently organized and remains easily accessible. A DBMS provides concurrency, security, data integrity, consistency, controls redundancy and data independence. In this project the Relational DBMS (RDBMS) used is MySQL. It is an open source software which uses SQL (Structured Query Language) which is a standard language for storing, manipulating and retrieving data in databases.

Java connections : To connect the database with the front end we use a java connector JDBC (Java Database Connectivity). JDBC is an application programming interface (API) for the programming language Java, which defines how a client may access a database. It is Java based data access technology and used for Java database connectivity. It is part of the Java Standard Edition platform, from Oracle Corporation. To achieve connectivity we use JSPs (Java Server Pages) and Servlets in this project. Java Server Pages (JSPs) is a technology that helps software developers create dynamically generated web pages based on HTML, XML, or other document types. 

Software Requirements: 
NetBeans  
MySQL Workbench
J2SE [J2EE], JDBC

Hardware Requirements:
Processor – Pentium IV or above
RAM – 2 GB or more
Hard disk – 3 GB or more